fix: add validation against repeated ports
